Fixed the example in the README.md
