+++ README.md
@@ -72,6 +72,7 @@ Code Example
     ```python
     import mdtraj as md
     import avtraj as avt
+    import json
     
     # First load an MD trajectory by mdtraj
     xtc_filename = './test/data/xtc/1am7_corrected.xtc'
@@ -86,41 +87,55 @@ Code Example
         'linker_length': 20.0,
         'linker_width': 1.0,
         'radius1': 3.5,
-        'simulation_grid_resolution': 0.5,
+        'simulation_grid_resolution': 1.0,
+        'residue_seq_number': 57,
+        'atom_name': 'CA'
     }
     
-    # Pass a trajectory to fps.AVTrajectory. This creates an object, which can be 
-    # accessed as a list. The objects within the "list" are accessible volumes  
+    # Pass a trajectory to fps.AVTrajectory. This creates an object, which can be
+    # accessed as a list. The objects within the "list" are accessible volumes
     av_traj_donor = avt.AVTrajectory(
         traj,
         av_parameters=av_parameters_donor,
-        name='57',
-        attachment_atom_selection='resSeq 57 and name CB'
+        name='57'
     )
     # These accessible volumes can be saved as xyz-file
-    av_traj_donor[0].save_av('57')
+    av_traj_donor[0].save_av()
     
     av_parameters_acceptor = {
         'simulation_type': 'AV1',
         'linker_length': 20.0,
         'linker_width': 1.0,
         'radius1': 3.5,
-        'simulation_grid_resolution': 0.5,
+        'simulation_grid_resolution': 1.0,
+        'residue_seq_number': 136,
+        'atom_name': 'CA'
     }
     
     # The dye parameters can either be passed explicitly on creation of the object
     av_traj_acceptor = avt.AVTrajectory(
-        traj, 
+        traj,
         av_parameters=av_parameters_acceptor,
-        name='136',
-        attachment_atom_selection='resSeq 136 and name CB'
+        name='136'
     )
+    av_traj_acceptor[0].save_av()
     
-    # To calculate a trajectory of distances and distance distributions first a 
+    
+    # To calculate a trajectory of distances and distance distributions first a
     # labeling file and a "distance file" needs to be specified. The distance
     # file contains a set of labeling positions and distances and should be
-    # compatible to the labeling files used by the software "Olga". By default the 
-    av_dist = avt.AvDistanceTrajectory(traj, './test/data/labeling.fps.json')
+    # compatible to the labeling files used by the software "Olga".
+    # or by the tool `label_structure` provided by the software ChiSurf.
+    labeling_file = './test/data/labeling.fps.json'
+    av_dist = avt.AvDistanceTrajectory(
+        traj,
+        json.load(
+            open(
+                labeling_file
+            )
+        )
+    )
+    print(av_dist[:10])
     ```
