+++ test.py
@@ -3,6 +3,8 @@ import avtraj as avt
 import json
 
 # First load an MD trajectory by mdtraj
+import avtraj.trajectory
+
 xtc_filename = './test/data/xtc/1am7_corrected.xtc'
 topology_filename = './test/data/xtc/1am7_protein.pdb'
 traj = md.load(
@@ -15,35 +17,39 @@ av_parameters_donor = {
     'linker_length': 20.0,
     'linker_width': 1.0,
     'radius1': 3.5,
-    'simulation_grid_resolution': 0.5,
+    'simulation_grid_resolution': 1.0,
+    'residue_seq_number': 57,
+    'atom_name': 'CA'
 }
 
 # Pass a trajectory to fps.AVTrajectory. This creates an object, which can be
 # accessed as a list. The objects within the "list" are accessible volumes
-av_traj_donor = avt.AVTrajectory(
+av_traj_donor = avtraj.trajectory.AVTrajectory(
     traj,
     av_parameters=av_parameters_donor,
-    name='57',
-    attachment_atom_selection='resSeq 57 and name CB'
+    name='57'
 )
 # These accessible volumes can be saved as xyz-file
-av_traj_donor[0].save_av('57')
+av_traj_donor[0].save_av()
 
 av_parameters_acceptor = {
     'simulation_type': 'AV1',
     'linker_length': 20.0,
     'linker_width': 1.0,
     'radius1': 3.5,
-    'simulation_grid_resolution': 0.5,
+    'simulation_grid_resolution': 1.0,
+    'residue_seq_number': 136,
+    'atom_name': 'CA'
 }
 
 # The dye parameters can either be passed explicitly on creation of the object
-av_traj_acceptor = avt.AVTrajectory(
+av_traj_acceptor = avtraj.trajectory.AVTrajectory(
     traj,
     av_parameters=av_parameters_acceptor,
-    name='136',
-    attachment_atom_selection='resSeq 136 and name CB'
+    name='136'
 )
+av_traj_acceptor[0].save_av()
+
 
 # To calculate a trajectory of distances and distance distributions first a
 # labeling file and a "distance file" needs to be specified. The distance
@@ -51,7 +57,7 @@ av_traj_acceptor = avt.AVTrajectory(
 # compatible to the labeling files used by the software "Olga".
 # or by the tool `label_structure` provided by the software ChiSurf.
 labeling_file = './test/data/labeling.fps.json'
-av_dist = avt.AvDistanceTrajectory(
+av_dist = avtraj.trajectory.AvDistanceTrajectory(
     traj,
     json.load(
         open(
@@ -59,4 +65,4 @@ av_dist = avt.AvDistanceTrajectory(
         )
     )
 )
-print(av_dist[0])
+print(av_dist[:10])
